+++ NEWS.md
@@ -1,5 +1,7 @@
 # dunlin 0.1.10.9000
 
+* Added `subject_level_flag()` function.
+
 # dunlin 0.1.10
 
 * The log printed by `reformat` when `verbose = TRUE` is limited to the tables and column existing in the reformatted data set.